src/requirements.txt
moviepy==1.0.3
Pillow==10.0.0
aeneas==1.7.3.0
numpy>=1.17.3
tqdm>=4.11.2
decorator>=4.0.2
lxml>=3.6.0
beautifulsoup4>=4.5.1
ffmpeg-python>=0.2.0

